# docs/swagger.yaml
basePath: /api/v1/
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API сервис - тестовое задание для стажировки в AvitoTech
  version: "1.0"
paths:
  /add_funds:
    post:
      consumes:
      - application/json
      description: add funds (sum) for user (id)
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Add Funds
  /funds_transfer:
    post:
      consumes:
      - application/json
      description: transfer funds (sum) from user (sender_id) to user (receiver_id)
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      - description: balance will convert from RUB to currency
        in: path
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Funds Transfer
  /get_balance:
    get:
      consumes:
      - application/json
      description: get user balance for user (id)
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Balance
  /write_off_funds:
    post:
      consumes:
      - application/json
      description: writes off funds (sum) for user (id)
      parameters:
      - description: input
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Write Off Funds
swagger: "2.0"